eventos_backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
